libctf: API review: delete ctf_ref
This function was meant to be used to let you bump the refcount on dicts
so that you could use them if you were passed an opened dict by someone
else who went on to close it.  The only instance of this in the libctf
API was ctf_archive_iter(), which has been deleted: and in any case,
ctf_archive_next() and therefore also ctf_archive_iter() these days
cache opened dicts, so it was entirely safe to just hang on to them
and keep using them after the ctf_archive_iter() iteration function
returned.

Useless: delete.