libctf, include, gdb: API review: ctf_get_arc -> ctf_dict_arc, endianness
libctf lets you specify the endianness of the symbol table, for those
cases where the symbol table is not derived from an ELF file.  It has
not one but two functions to do this, ctf_symsect_endianness and
ctf_symsect_endianness.

This is such an obscure use case that having two functions seems like
overkill.  We can reduce it to only one by saying that ctf_get_arc
can synthesise an archive if one doesn't already exist: then all dual
ctf_* and ctf_arc_* functions can be reduced to just the ctf_arc_*
form.  While we're at it, rename ctf_get_arc to ctf_dict_arc, like all
other functions that take a dict and return something else.

We do need to add a parameter to ctf_dict_arc, to allow for one use
case. Normally, if you call it and it synthesises an archive for you,
you don't want to be bothered with it: ctf_dict_close frees it
automatically and you can ignore its lifetime. But if you are
specifically calling ctf_dict_arc because you know you opened the
archive yourself (it's not a synthetic one) and you need to close
it explicitly, you absolutely do not want one to be synthesised: this
will lead to a double-free when you do a ctf_dict_close() followed by a
ctf_close().

So add a "do not synthesise an archive, just return the real thing"
parameter to ctf_dict_arc accordingly.