libctf: API review: C++ compatibility for enumized flags
A number of the recently-enumized flags are in flags for which we want
the caller to be able to specify "no flags set" by just passing 0.

This is, alas, not C++-compatible: C++ only defines implicit conversions
from enum to int, not the other way round, not even for 0.  So make them
all typedefs to int on C++.  (Does not affect name mangling: this is all
in extern "C". It does mean you'll get fewer warnings in C++ code than
in C code because the compiler can no longer do enum coverage analysis,
but in most cases this is harmless, except for ctf_type_kind_t.)